+++ src/main/java/ai/starlake/transpiler/redshift/RedshiftExpressionTranspiler.java
@@ -1,0 +1,129 @@
+/**
+ * Starlake.AI JSQLTranspiler is a SQL to DuckDB Transpiler.
+ * Copyright (C) 2024 Andreas Reichel <andreas@manticore-projects.com> on behalf of Starlake.AI
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package ai.starlake.transpiler.redshift;
+
+import ai.starlake.transpiler.JSQLExpressionTranspiler;
+import ai.starlake.transpiler.JSQLTranspiler;
+import net.sf.jsqlparser.expression.AnalyticExpression;
+import net.sf.jsqlparser.expression.CaseExpression;
+import net.sf.jsqlparser.expression.Expression;
+import net.sf.jsqlparser.expression.Function;
+import net.sf.jsqlparser.expression.LongValue;
+import net.sf.jsqlparser.expression.WhenClause;
+import net.sf.jsqlparser.expression.operators.relational.ExpressionList;
+import net.sf.jsqlparser.expression.operators.relational.GreaterThan;
+import net.sf.jsqlparser.expression.operators.relational.MinorThan;
+
+public class RedshiftExpressionTranspiler extends JSQLExpressionTranspiler {
+  public RedshiftExpressionTranspiler(JSQLTranspiler transpiler, StringBuilder buffer) {
+    super(transpiler, buffer);
+  }
+
+  enum TranspiledFunction {
+    // @FORMATTER:OFF
+    BPCHARCMP;
+    // @FORMATTER:ON
+
+
+    @SuppressWarnings({"PMD.EmptyCatchBlock"})
+    public static TranspiledFunction from(String name) {
+      TranspiledFunction function = null;
+      try {
+        function = Enum.valueOf(TranspiledFunction.class, name.toUpperCase());
+      } catch (Exception ignore) {
+        // nothing to do here
+      }
+      return function;
+    }
+
+    public static TranspiledFunction from(Function f) {
+      return from(f.getName());
+    }
+  }
+
+  enum UnsupportedFunction {
+    ASINH, ACOSH, COSH, SINH, COTH, COSINE_DISTANCE, CSC, CSCH, EUCLIDEAN_DISTANCE, SEC, SECH
+
+    , APPROX_QUANTILES, APPROX_TOP_COUNT, APPROX_TOP_SUM
+
+    , SEARCH, VECTOR_SEARCH, APPENDS, EXTERNAL_OBJECT_TRANSFORM, GAP_FILL
+
+    ;
+
+    @SuppressWarnings({"PMD.EmptyCatchBlock"})
+    public static UnsupportedFunction from(String name) {
+      UnsupportedFunction function = null;
+      try {
+        function = Enum.valueOf(UnsupportedFunction.class, name.toUpperCase());
+      } catch (Exception ignore) {
+        // nothing to do here
+      }
+      return function;
+    }
+
+    public static UnsupportedFunction from(Function f) {
+      return from(f.getName());
+    }
+
+    public static UnsupportedFunction from(AnalyticExpression f) {
+      return from(f.getName());
+    }
+  }
+
+
+  @SuppressWarnings({"PMD.CyclomaticComplexity", "PMD.ExcessiveMethodLength"})
+  public void visit(Function function) {
+    String functionName = function.getName();
+
+    if (UnsupportedFunction.from(function) != null) {
+      throw new RuntimeException(
+          "Unsupported: " + functionName + " is not supported by DuckDB (yet).");
+    } else if (functionName.endsWith("$$")) {
+      // work around for transpiling already transpiled functions twice
+      // @todo: figure out a better way to achieve that
+      function.setName(functionName.substring(0, functionName.length() - 2));
+      super.visit(function);
+      return;
+    }
+
+    if (function.getMultipartName().size() > 1
+        && function.getMultipartName().get(0).equalsIgnoreCase("SAFE")) {
+      warning("SAFE prefix is not supported.");
+      function.getMultipartName().remove(0);
+    }
+
+    Expression rewrittenExpression = null;
+    ExpressionList<?> parameters = function.getParameters();
+    TranspiledFunction f = TranspiledFunction.from(functionName);
+    if (f != null) {
+      switch (f) {
+        case BPCHARCMP:
+          rewrittenExpression = new CaseExpression(new LongValue(0),
+              new WhenClause(new GreaterThan(parameters.get(0), parameters.get(1)),
+                  new LongValue(1)),
+              new WhenClause(new MinorThan(parameters.get(0), parameters.get(1)),
+                  new LongValue(-1)));
+          break;
+      }
+    }
+    if (rewrittenExpression == null) {
+      super.visit(function);
+    } else {
+      rewrittenExpression.accept(this);
+    }
+  }
+}