+++ src/test/java/ai/starlake/transpiler/CLITest.java
@@ -18,7 +18,7 @@ public class CLITest extends JSQLTranspilerTest {
   @ParameterizedTest(name = "{index} {0} {1}: {2}")
   @MethodSource("getSqlTestMap")
   @Disabled
-  void transpile(File f, int idx, SQLTest t) throws Exception {
+  protected void transpile(File f, int idx, SQLTest t) throws Exception {
 
   }