+++ src/test/java/ai/starlake/transpiler/redshift/RedshiftTranspilerTest.java
@@ -1,5 +1,7 @@
-package ai.starlake.transpiler;
+package ai.starlake.transpiler.redshift;
 
+import ai.starlake.transpiler.JSQLTranspiler;
+import ai.starlake.transpiler.JSQLTranspilerTest;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -7,9 +9,9 @@ import org.junit.jupiter.params.provider.MethodSource;
 import java.io.File;
 import java.util.stream.Stream;
 
-public class AmazonRedshiftTranspilerTest extends JSQLTranspilerTest {
+public class RedshiftTranspilerTest extends JSQLTranspilerTest {
   public final static String TEST_FOLDER_STR =
-      "build/resources/test/ai/starlake/transpiler/amazon_redshift";
+      "build/resources/test/ai/starlake/transpiler/redshift";
 
   static Stream<Arguments> getSqlTestMap() {
     return unrollParameterMap(getSqlTestMap(new File(TEST_FOLDER_STR).listFiles(FILENAME_FILTER),
@@ -18,7 +20,7 @@ public class AmazonRedshiftTranspilerTest extends JSQLTranspilerTest {
 
   @ParameterizedTest(name = "{index} {0} {1}: {2}")
   @MethodSource("getSqlTestMap")
-  void transpile(File f, int idx, SQLTest t) throws Exception {
+  protected void transpile(File f, int idx, SQLTest t) throws Exception {
     super.transpile(f, idx, t);
   }